BackgroundRemover/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
pillow>=10.0.0
opencv-contrib-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# === AI & Advanced Features ===
rembg>=2.0.50
onnxruntime>=1.16.0

